# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  inventory.AdjustStockRequest:
    properties:
      delta:
        description: Can be negative (deductions) or positive (restocks)
        example: 25
        type: integer
      warehouse_id:
        example: 8f076135-d858-45ec-b9cc-0320df4ee99c
        type: string
    type: object
  inventory.CreateProductRequest:
    properties:
      name:
        example: Ergonomic Office Chair Premium
        type: string
      sku:
        example: PROD-XYZ-001
        type: string
    type: object
  inventory.CreateWarehouseRequest:
    properties:
      name:
        example: Lagos Central Distribution Hub
        type: string
    type: object
  inventory.ProductResponse:
    properties:
      name:
        example: Ergonomic Office Chair Premium
        type: string
      product_id:
        example: bc731ee0-3333-4f9e-a89c-a1d257211140
        type: string
      sku:
        example: PROD-XYZ-001
        type: string
    type: object
  inventory.StockLevelResponse:
    properties:
      product_id:
        example: bc731ee0-3333-4f9e-a89c-a1d257211140
        type: string
      quantity:
        example: 142
        type: integer
      reorder_point:
        example: 15
        type: integer
      warehouse_id:
        example: 8f076135-d858-45ec-b9cc-0320df4ee99c
        type: string
    type: object
  inventory.WarehouseResponse:
    properties:
      name:
        example: Lagos Central Distribution Hub
        type: string
      warehouse_id:
        example: 8f076135-d858-45ec-b9cc-0320df4ee99c
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant REP system for OneDesk.
  title: OneDesk API
  version: "1.0"
paths:
...
  /v1/products:
    get:
      description: Provides full product catalog metrics visibility owned by the matching
        business group profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.ProductResponse'
            type: array
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Core relational lookup tracking query fault context
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List catalog tracking records
      tags:
      - inventory
    post:
      consumes:
      - application/json
      description: Adds a unique tracking product variant profile under the current
        client scope. Enforces item ceiling controls.
      parameters:
      - description: Product registration data template
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inventory.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.ProductResponse'
        "400":
          description: Invalid structural input payload values
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Product catalog resource entitlement quota exhausted
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: SKU code entry collision matching active entity records
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a product item
      tags:
      - inventory
  /v1/products/{productId}/stock:
    get:
      description: Returns current stock level breakdowns and reorder triggers across
        all operational warehouse points.
      parameters:
      - description: Target inventory resource look up track key ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.StockLevelResponse'
            type: array
        "400":
          description: Malformed route identification parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No item registration matches the requested ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal query execution layer exception structural crash
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get product stock levels
      tags:
      - inventory
  /v1/products/{productId}/stock/adjust:
    post:
      consumes:
      - application/json
      description: Mutates available balance counts up or down across distinct warehouses.
        Rejects operations that result in a negative total.
      parameters:
      - description: Target catalog item system resource ID (UUID)
        in: path
        name: productId
        required: true
        type: string
      - description: Storage destination map and balancing metrics delta values
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inventory.AdjustStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.StockLevelResponse'
        "400":
          description: Malformed route parameters or tracking payload errors
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Target unique identification index resource missing
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Balancing step rejected because final allocation would fall
            below zero
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Adjust product stock level
      tags:
      - inventory
...
  /v1/warehouses:
    get:
      description: Returns a full list of storage locations associated with the active
        tenant.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventory.WarehouseResponse'
            type: array
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal service storage lookup engine failure
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all warehouses
      tags:
      - inventory
    post:
      consumes:
      - application/json
      description: Deploys a distinct inventory tracking location workspace for the
        tenant.
      parameters:
      - description: Warehouse creation parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inventory.CreateWarehouseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.WarehouseResponse'
        "400":
          description: Invalid payload or request processing error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new warehouse
      tags:
      - inventory
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"